Split out runtime
